+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - SelectMenu component
+- Icon Button component
 
 ## 1.2.5 - 2022-12-16
 
@@ -26,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.2.3 - 2022-12-15
 
-## 1.2.2 - 2022-12-15
+- Icon button
 
 ## 1.2.1 - 2022-12-15